# texlive/combine: avoid sorting on store paths (fix CA stability)