In https://github.com/NixOS/nix/issues/5333 I noticed that adding
a minor change to `bash` derivation triggers rebuilds in CA for
`bash` (expected) and `texlive.combine` (unexpected).

regnat debugged derivation instability down to sort on outPath.

The change avoid sorting on outPaths and uses pname of the derivations.

Tested on `R` derivation.

Closes: https://github.com/NixOS/nix/issues/5333