@veprbl 2021-10-06
Thank you for looking into this. I've noticed you are exchanging order of `lib.sort` and `lib.unique`. I would assume that the original ordering produces a better evaluation performance. Did you look at that before implementing the change?