@xaverdh 2021-10-06
> Thank you for looking into this. I've noticed you are exchanging order of `lib.sort` and `lib.unique`. I would assume that the original ordering produces a better evaluation performance. Did you look at that before implementing the change?

https://github.com/NixOS/nixpkgs/commit/85605c8a296bd54e9c12e1bd1c33bf1e100d40b7 and https://github.com/NixOS/nixpkgs/commit/ce84cc06bf3a6311c67e73b6675fc5c1317dec23 for some discussion on this