@trofi 2021-10-06
Oh, I did not take performance change into account assuming sorting by `pname` should not be heavyweight.  Looks like it's now ~50% slower. I'll have a look.

Here is my naive benchmark to build a single texlive package out of all available packages:

```
# before
$ time nix-instantiate -E 'with import ./. {}; let ps = lib.filterAttrs (n: v: lib.isAttrs v && lib.hasAttr "pkgs" v) texlive; in texlive.combine ps'

real    0m2,206s
user    0m1,776s
sys     0m0,291s

# after
$ time nix-instantiate -E 'with import ./. {}; let ps = lib.filterAttrs (n: v: lib.isAttrs v && lib.hasAttr "pkgs" v) texlive; in texlive.combine ps'

real    0m3,151s
user    0m2,877s
sys     0m0,306s
```