@xworld21 2021-10-06
Sorting by pname is a good idea. In principle also by tlType, but if lib.sort is stable, the order should be deterministic no matter what.

Re performance, wouldn't `genericClosure` be a lot faster?