@trofi 2021-10-07
Looked at performance of `mkUniqueOutPaths = pkgs: lib.unique (map (p: p.outPath) (sortedByPN (builtins.filter lib.isDerivation pkgs)));` a bit more. For all-of-tex benchmark `filter` returns 63K entries, `unique` returns 4K entries. Both are mostly busy comparing string values (derivation paths or package names):

```
    27.32%  nix-instantiate  libc-2.33.so         [.] __strcmp_avx2
    11.25%  nix-instantiate  libnixexpr.so        [.] nix::EvalState::eqValues
     4.02%  nix-instantiate  libnixexpr.so        [.] nix::ExprSelect::eval
     2.78%  nix-instantiate  libgc.so.1.4.3       [.] GC_mark_from
     2.31%  nix-instantiate  libnixexpr.so        [.] nix::ExprVar::eval
     2.00%  nix-instantiate  libnixexpr.so        [.] nix::EvalState::forceValue
```

Before this change `unique` did indeed decreased input for `sort` 16x. Now it does not.

Not sure what I could do to get the speed back. I tried a few simple tricks, but most of them make things worse. The fact that `sort` has to work on indirection via comparison does not help either (but maybe it's a minor part of overhead).