@xaverdh 2021-10-07
Can we assume that the mapping pname -> outPath is injective, i.e. there are no two distinct packages with the same pname?
Then I can imagine
```nix
let
  sortUnique =
    let hasLast = x: xs: if xs != [] && lib.last xs == x then true else false;
    in f: ys: lib.foldl' (xs: x: if hasLast x xs then xs else xs ++ [x]) [] (lib.sort f ys);

  sortedByPNUnique = sortUnique (a: b: a.pname < b.pname);
  sortedUniqueStrings = sortUnique (a: b: a < b);

  mkUniqueOutPaths = pkgs: lib.unique
    (map (p: p.outPath) (sortedByPNUnique (builtins.filter lib.isDerivation pkgs)));
in [...]
```
to have better runtime..