@xworld21 2021-10-07
After looking at this, I am pretty confident that there is no need for sort in `mkUniqueOutPaths`. It may be an historical artefact from the older `fastUnique` (which looks similar to @xaverdh's `sortUnique` by the way). On the other hand, `formatPNames` and `hyphenPNames` *must* be sorted, but those lists are short (<70 elements). @veprbl do you agree? I suggest moving `lib.sort` inside `formatPNames`, `hyphenPNames` and run the tests.

> Can we assume that the mapping pname -> outPath is injective

No, each `pname` may have different variants (`doc`, `src'), but you can assume that `name` -> `outPath` is injective.